[Week2-step3] : step2 파일 생성 및 이동
